requirements.txt
aiohttp-3.8.4
python-dotenv
SQLAlchemy==1.4.48